Refactored debug
